Removing UART testing - swithcing to STM Monitor

+++ PWM_output/Core/Inc/adc.h
@@ -23,7 +23,7 @@ typedef struct {
 } adc_curr_raw_t;
 
 /*Use privately the pointer for hadc*/
-void ADC_Init(ADC_HandleTypeDef *hadc, uint16_t theta_0);
+void ADC_Init(ADC_HandleTypeDef *hadc, float theta_0);
 
 
 void DWT_Init(void);
